+++ examples/MOTD/run.py
@@ -7,12 +7,6 @@ import logging
 # Enthought library imports.
 from envisage.api import Application, CorePlugin
 
-# Example plugins.
-from acme.motd.motd_plugin import MOTDPlugin
-from acme.motd.software_quotes.software_quotes_plugin import (
-    SoftwareQuotesPlugin,
-)
-
 
 # Do whatever you want to do with log messages! Here we create a log file.
 logger = logging.getLogger()
@@ -22,7 +16,12 @@ logger.setLevel(logging.DEBUG)
 
 def main():
     """ Run the application. """
-
+    # Import here so that this script can be run from anywhere without
+    # having to install the packages.
+    from acme.motd.motd_plugin import MOTDPlugin
+    from acme.motd.software_quotes.software_quotes_plugin import (
+        SoftwareQuotesPlugin,
+    )
     # Create an application containing the appropriate plugins.
     application = Application(
         id="acme.motd",
@@ -34,4 +33,9 @@ def main():
 
 
 if __name__ == "__main__":
-    main()
+    # This context manager is added so that one can run this example from any
+    # directory without necessarily having installed the examples as packages.
+    from envisage.examples._demo import demo_path
+
+    with demo_path(__file__):
+        main()
